tas,libtas: use cycles for poll cycle delays
This should guarantee that both sides are measuring the same way

+++ tas/slow/appif.h
@@ -52,7 +52,7 @@ struct app_context {
   struct app_doorbell *doorbell;
 
   int ready, evfd;
-  uint32_t last_ts;
+  uint64_t last_ts;
   struct app_context *next;
 
   struct {
